requirements.txt
flask>=2.0.1
flask-wtf>=1.0.0
flask-session>=0.4.0
pandas>=1.5.3
openai>=1.6.0
anthropic>=0.8.0
biopython>=1.81
requests>=2.31.0
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.1
openpyxl>=3.1.2
gunicorn>=20.1.0  # For production deployment
tenacity>=8.0.0  # For retry logic and rate limiting
python-dotenv>=1.0.0  # For .env file support
